requirements.txt
astrapy
sentence-transformers
python-dotenv
requests
beautifulsoup4
lxml
playwright
fastapi
uvicorn